rag-adv/requirements.txt
numpy
streamlit
langchain-community
PyPDF2
unstructured
nltk
sentence-transformers
langchain-huggingface
faiss-cpu
python-docx
